requirements.txt
requests==2.32.3
pytesseract==0.3.13
aiogram==3.17.0
openai==1.61.1
python-dotenv==1.0.1
